Refactored edge service working
